third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-focus-style-expected.txt

Check focus style for multiple fields time input UI

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "-webkit-datetime-edit-hour-field"; style.backgroundColor is highlight
PASS "-webkit-datetime-edit-hour-field"; style.color is highlighttext
PASS "-webkit-datetime-edit-minute-field"; style.backgroundColor is highlight
PASS "-webkit-datetime-edit-minute-field"; style.color is highlighttext
PASS "-webkit-datetime-edit-second-field"; style.backgroundColor is highlight
PASS "-webkit-datetime-edit-second-field"; style.color is highlighttext
PASS "-webkit-datetime-edit-millisecond-field"; style.backgroundColor is highlight
PASS "-webkit-datetime-edit-millisecond-field"; style.color is highlighttext
PASS "-webkit-datetime-edit-ampm-field"; style.backgroundColor is highlight
PASS "-webkit-datetime-edit-ampm-field"; style.color is highlighttext


PASS successfullyParsed is true

TEST COMPLETE

Please run this test within DumpRenderTree.
